Fix the keyboard handler hook to allow multiple instances
Instead of setting `document.onkeyup = ...`, now it's using
`document.addEventListener('keyup', ...)`.

That way we can add multiple event listeners, which will come in handy
with arrow drawing keyboard events.
