+++ src/pages/AnnotationPage.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 import { Provider } from 'react-redux'
 import { PersistGate } from 'redux-persist/integration/react'
 import Code from '../components/Code'
@@ -6,7 +6,9 @@ import Controls from '../components/Controls'
 import useCode from '../hooks/useCode'
 import useKeyboardHandler from '../hooks/useKeyboardHandler'
 import useSource from '../hooks/useSource'
-import createStore from '../store'
+import { clearSelection, removeArrow, removeMarker } from '../reducer'
+import createStore, { useDispatch, useSelector } from '../store'
+import { redo, undo } from '../undoable'
 
 export default function AnnotationPage() {
   const source = useSource()
@@ -43,10 +45,7 @@ function LoadingPage({ error }: { error: Error | null }) {
 }
 
 function LoadedPage({ code }: { code: string }) {
-  const handler = useKeyboardHandler()
-  useEffect(() => {
-    document.onkeydown = handler
-  }, [handler])
+  useReducerKeyboardHandler()
 
   return (
     <div>
@@ -56,4 +55,52 @@ function LoadedPage({ code }: { code: string }) {
       </div>
     </div>
   )
+}
+
+function useReducerKeyboardHandler() {
+  const dispatch = useDispatch()
+  const currentSelection = useSelector((state) => state.currentSelection)
+
+  const handler = React.useCallback(
+    (event: KeyboardEvent) => {
+      const isMetaOn = event.ctrlKey || event.metaKey
+      if (event.key === 'z' && isMetaOn && event.shiftKey) {
+        // redo
+        event.preventDefault()
+        event.stopPropagation()
+        dispatch(redo())
+      } else if (event.key === 'z' && isMetaOn) {
+        // undo
+        event.preventDefault()
+        event.stopPropagation()
+        dispatch(undo())
+      } else if (
+        ['Backspace', 'Delete'].includes(event.key) &&
+        currentSelection
+      ) {
+        // remove currently selected marker/arrow
+        switch (currentSelection.type) {
+          case 'text': {
+            return
+          }
+          case 'marker': {
+            dispatch(removeMarker(currentSelection.marker))
+            return
+          }
+          case 'arrow': {
+            dispatch(removeArrow(currentSelection.arrow))
+            return
+          }
+        }
+      } else if (event.key === 'Escape' && currentSelection) {
+        dispatch(clearSelection())
+        if (currentSelection.type === 'text') {
+          document.getSelection()?.removeAllRanges()
+        }
+      }
+    },
+    [dispatch, currentSelection],
+  )
+
+  useKeyboardHandler(handler)
 }
